requirements.txt
colorama==0.4.6
mss==10.0.0
numpy==2.2.4
pillow==11.2.1
PyYAML==6.0.2
